requirements.txt
torch
torchvision
pyyaml
fastapi
uvicorn
python-multipart
pillow
numpy